requirements.txt
# Core framework
fastapi==0.121.1
uvicorn==0.38.0

# Google Sheets integration
gspread==6.2.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3

# Image processing
pillow==12.0.0

# HTTP requests
requests==2.32.3

# Environment variables
python-dotenv==1.2.1

# Gemini API client
# gemini-vision==0.3.2
google-generativeai==0.3.2


# Logging, utilities, etc.
typing-extensions==4.15.0
annotated-types==0.7.0
annotated-doc==0.0.4
anyio==4.11.0
pydantic==2.12.4
pydantic-core==2.41.5
sniffio==1.3.1
click==8.3.0
colorama==0.4.6
h11==0.16.0


beautifulsoup4
lxml
python-amazon-paapi


